Cleaned RHC Benchmark Cohort

The cleaned Right Heart Catheterization (RHC) cohort from the SUPPORT study, as produced by prepare_rhc_data. RHC was associated with harm in the original analysis (Connors et al. 1996); the data are bundled so the validation vignette and tests run offline and deterministically.

Usage

rhc_sample

Format

A list with two elements:

data

A data frame with 5735 rows and 63 columns. swang1 is the binary exposure (1 = RHC); dth30 is 30-day mortality and los is length of stay in days.

covariates

A character vector of the 50 confounder column names used by the adjustment methods.

Details

The raw file's quirks are resolved here: the write.csv row-names column is dropped, the cat2 category "None" (stored as the literal text "NA") is restored, remaining literal "NA" strings become real missing values, the swang1 exposure and dth30/death outcomes are recoded 0/1, and los is derived from the admission and discharge dates. The heavy-missingness columns adld3p and urin1 are excluded from the covariate vector.

Examples

data(rhc_sample)
str(rhc_sample$data)
#> 'data.frame':	5735 obs. of  63 variables:
#>  $ cat1    : chr  "COPD" "MOSF w/Sepsis" "MOSF w/Malignancy" "ARF" ...
#>  $ cat2    : chr  "None" "None" "MOSF w/Sepsis" "None" ...
#>  $ ca      : chr  "Yes" "No" "Yes" "No" ...
#>  $ sadmdte : num  11142 11799 12083 11146 12035 ...
#>  $ dschdte : num  11151 11844 12143 11183 12037 ...
#>  $ dthdte  : num  NA 11844 NA 11183 12037 ...
#>  $ lstctdte: num  11382 11844 12400 11182 12036 ...
#>  $ death   : num  0 1 0 1 1 0 0 1 0 0 ...
#>  $ cardiohx: num  0 1 0 0 0 0 0 0 0 0 ...
#>  $ chfhx   : num  0 1 0 0 0 1 0 0 0 0 ...
#>  $ dementhx: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ psychhx : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ chrpulhx: num  1 0 0 0 0 1 0 0 0 0 ...
#>  $ renalhx : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ liverhx : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ gibledhx: num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ malighx : num  1 0 1 0 0 0 1 0 0 1 ...
#>  $ immunhx : num  0 1 1 1 0 0 0 0 0 0 ...
#>  $ transhx : num  0 1 0 0 0 0 0 1 0 0 ...
#>  $ amihx   : num  0 0 0 0 0 0 0 0 0 0 ...
#>  $ age     : num  70.3 78.2 46.1 75.3 67.9 ...
#>  $ sex     : chr  "Male" "Female" "Female" "Female" ...
#>  $ edu     : num  12 12 14.07 9 9.95 ...
#>  $ surv2md1: num  0.641 0.755 0.317 0.441 0.437 ...
#>  $ das2d3pc: num  23.5 14.8 18.1 22.9 21.1 ...
#>  $ t3d30   : num  30 30 30 30 2 30 30 30 30 30 ...
#>  $ dth30   : num  0 0 0 0 1 0 0 0 0 0 ...
#>  $ aps1    : num  46 50 82 48 72 38 29 25 47 48 ...
#>  $ scoma1  : num  0 0 0 0 41 0 26 100 0 0 ...
#>  $ meanbp1 : num  41 63 57 55 65 115 67 128 53 73 ...
#>  $ wblc1   : num  22.1 28.9 0.05 23.3 29.7 ...
#>  $ hrt1    : num  124 137 130 58 125 134 135 102 118 141 ...
#>  $ resp1   : num  10 38 40 26 27 36 10 34 30 40 ...
#>  $ temp1   : num  38.7 38.9 36.4 35.8 34.8 ...
#>  $ pafi1   : num  68 218 276 157 478 ...
#>  $ alb1    : num  3.5 2.6 3.5 3.5 3.5 ...
#>  $ hema1   : num  58 32.5 21.1 26.3 24 ...
#>  $ bili1   : num  1.01 0.7 1.01 0.4 1.01 ...
#>  $ crea1   : num  1.2 0.6 2.6 1.7 3.6 ...
#>  $ sod1    : num  145 137 146 117 126 138 136 136 136 146 ...
#>  $ pot1    : num  4 3.3 2.9 5.8 5.8 ...
#>  $ paco21  : num  40 34 16 30 17 68 45 26 40 30 ...
#>  $ ph1     : num  7.36 7.33 7.36 7.46 7.23 ...
#>  $ swang1  : num  0 1 1 0 1 0 0 0 0 1 ...
#>  $ wtkilo1 : num  64.7 45.7 0 54.6 78.4 ...
#>  $ dnr1    : chr  "No" "No" "No" "No" ...
#>  $ ninsclas: chr  "Medicare" "Private & Medicare" "Private" "Private & Medicare" ...
#>  $ resp    : chr  "Yes" "No" "No" "Yes" ...
#>  $ card    : chr  "Yes" "No" "Yes" "No" ...
#>  $ neuro   : chr  "No" "No" "No" "No" ...
#>  $ gastr   : chr  "No" "No" "No" "No" ...
#>  $ renal   : chr  "No" "No" "No" "No" ...
#>  $ meta    : chr  "No" "No" "No" "No" ...
#>  $ hema    : chr  "No" "No" "No" "No" ...
#>  $ seps    : chr  "No" "Yes" "No" "No" ...
#>  $ trauma  : chr  "No" "No" "No" "No" ...
#>  $ ortho   : chr  "No" "No" "No" "No" ...
#>  $ adld3p  : num  0 NA NA NA NA 0 NA NA NA NA ...
#>  $ urin1   : num  NA 1437 599 NA 64 ...
#>  $ race    : chr  "white" "white" "white" "white" ...
#>  $ income  : chr  "Under $11k" "Under $11k" "$25-$50k" "$11-$25k" ...
#>  $ ptid    : chr  "00005" "00007" "00009" "00010" ...
#>  $ los     : num  9 45 60 37 2 7 42 34 11 19 ...
length(rhc_sample$covariates)
#> [1] 50